+++ lib/ts/recipe/webauthn/index.ts
@@ -155,6 +155,7 @@ export default class RecipeWrapper {
     static signUp(input: {
         webauthnGeneratedOptionsId: string;
         credential: RegistrationResponseJSON;
+        shouldTryLinkingWithSessionUser?: boolean;
         options?: RecipeFunctionOptions;
         userContext: any;
     }): Promise<
@@ -197,6 +198,7 @@ export default class RecipeWrapper {
     static signIn(input: {
         webauthnGeneratedOptionsId: string;
         credential: AuthenticationResponseJSON;
+        shouldTryLinkingWithSessionUser?: boolean;
         options?: RecipeFunctionOptions;
         userContext: any;
     }): Promise<
@@ -376,6 +378,7 @@ export default class RecipeWrapper {
      */
     static registerCredentialWithSignUp(input: {
         email: string;
+        shouldTryLinkingWithSessionUser?: boolean;
         options?: RecipeFunctionOptions;
         userContext: any;
     }): Promise<
@@ -427,7 +430,11 @@ export default class RecipeWrapper {
      *
      * @returns `{ status: "OK", ...}` if successful along a description of the user details (id, etc.) and email
      */
-    static authenticateCredentialWithSignIn(input: { options?: RecipeFunctionOptions; userContext: any }): Promise<
+    static authenticateCredentialWithSignIn(input: {
+        options?: RecipeFunctionOptions;
+        userContext: any;
+        shouldTryLinkingWithSessionUser?: boolean;
+    }): Promise<
         | {
               status: "OK";
               user: User;
@@ -527,12 +534,12 @@ export default class RecipeWrapper {
               fetchResponse: Response;
           }
         | GeneralErrorResponse
-        | { status: "REGISTER_CREDENTIAL_NOT_ALLOWED"; reason: string }
+        | { status: "REGISTER_CREDENTIAL_NOT_ALLOWED"; reason?: string }
         | { status: "INVALID_EMAIL_ERROR"; err: string }
         | { status: "INVALID_CREDENTIALS_ERROR" }
         | { status: "OPTIONS_NOT_FOUND_ERROR" }
         | { status: "INVALID_OPTIONS_ERROR" }
-        | { status: "INVALID_AUTHENTICATOR_ERROR"; reason: string }
+        | { status: "INVALID_AUTHENTICATOR_ERROR"; reason?: string }
         | { status: "AUTHENTICATOR_ALREADY_REGISTERED" }
         | { status: "FAILED_TO_REGISTER_USER"; error: any }
         | { status: "WEBAUTHN_NOT_SUPPORTED"; error: any }
@@ -614,11 +621,11 @@ export default class RecipeWrapper {
     }): Promise<
         | { status: "OK" }
         | GeneralErrorResponse
-        | { status: "REGISTER_CREDENTIAL_NOT_ALLOWED"; reason: string }
+        | { status: "REGISTER_CREDENTIAL_NOT_ALLOWED"; reason?: string }
         | { status: "INVALID_CREDENTIALS_ERROR" }
         | { status: "OPTIONS_NOT_FOUND_ERROR" }
         | { status: "INVALID_OPTIONS_ERROR" }
-        | { status: "INVALID_AUTHENTICATOR_ERROR"; reason: string }
+        | { status: "INVALID_AUTHENTICATOR_ERROR"; reason?: string }
     > {
         return Recipe.getInstanceOrThrow().recipeImplementation.registerCredential(input);
     }